Added two missing types in density_combine

+++ gmrflib/density.h
@@ -146,9 +146,9 @@ __BEGIN_DECLS
    2 \phi(x) \Phi(alpha*x)
 */
 typedef struct {
-	float xi;
-	float omega;
-	float alpha;
+	double xi;
+	double omega;
+	double alpha;
 } GMRFLib_sn_param_tp;
 
 typedef struct {
@@ -171,32 +171,32 @@ typedef struct {
 	/**
 	 * \brief The mean in standarised scale.
 	 */
-	float mean;
+	double mean;
 
 	/**
 	 * \brief The standard deviation in standardised scale.
 	 */
-	float stdev;
+	double stdev;
 
 	/**
 	 * \brief The standardised skewness
 	 */
-	float skewness;
+	double skewness;
 
 	/**
 	 * \brief The mean in users own scale (not standardised).
 	 */
-	float user_mean;
+	double user_mean;
 
 	/**
 	 * \brief The standard deviation in users own scale (not standardised).
 	 */
-	float user_stdev;
+	double user_stdev;
 
 	/** 
 	 * \brief The mode in user scale (set if !NAN)
 	 */
-	float user_mode;
+	double user_mode;
 
 	/**
 	 * \brief The offset that is used for standarisation.
@@ -204,7 +204,7 @@ typedef struct {
 	 * The density is for the random variable (x-m)/sd where \c m is \c GMRFLib_density_tp::std_mean and \c sd is
 	 * GMRFLib_density_tp::std_stdev.
 	 */
-	float std_mean;
+	double std_mean;
 
 	/** 
 	 * \brief The scale that is used for standarisation.
@@ -212,7 +212,7 @@ typedef struct {
 	 * The density is for the random variable (x-m)/sd where \c m is \c GMRFLib_density_tp::std_mean and \c sd is
 	 * GMRFLib_density_tp::std_stdev.
 	 */
-	float std_stdev;				       /* this is the the standarisation */
+	double std_stdev;				       /* this is the the standarisation */
 
 	/*
 	 * The rest of the paramers is for internal use only. 
@@ -223,13 +223,13 @@ typedef struct {
 	 */
 	GMRFLib_density_type_tp type;
 
-	float x_min, x_max;				       /* range for the log_correction, ALSO used by the others */
+	double x_min, x_max;				       /* range for the log_correction, ALSO used by the others */
 
 	// /*
 	// * params for the GMRFLib_DENSITY_TYPE_GAUSSIAN 
 	// */
-	// float mean_gaussian; // use now mean and stdev
-	// float stdev_gaussian; // use now mean and stdev
+	// double mean_gaussian; // use now mean and stdev
+	// double stdev_gaussian; // use now mean and stdev
 
 	/*
 	 * params for the GMRFLib_DENSITY_TYPE_SKEWNORMAL 
@@ -239,7 +239,7 @@ typedef struct {
 	/*
 	 * params for the GMRFLib_DENSITY_TYPE_SCGAUSSIAN 
 	 */
-	float log_norm_const;				       /* log(norm_const), divide by norm_const to get the normalised density.  */
+	double log_norm_const;				       /* log(norm_const), divide by norm_const to get the normalised density.  */
 	GMRFLib_spline_tp *log_correction;
 	GMRFLib_spline_tp *P;
 	GMRFLib_spline_tp *Pinv;
